Update Role.java
Constants are variables whose value does not change during the runtime of a program after initialization. Oftentimes, constants are used in multiple locations across different subroutines.
It is important that the names of constants follow a consistent and easily recognizable pattern. This way, readers immediately understand that the referenced value does not change, which simplifies debugging.
Or, in the case of primitive constants, that accessing the constant is thread-safe.
This rule checks that all constant names match a given regular expression.